requirements.txt
streamlit>=1.31.0
phidata>=2.4.0
groq>=0.4.0
python-dotenv>=1.0.0
serpapi==0.1.5
google-search-results==2.4.2